Modify the script and add SDK link empty judgment

+++ .github/ci/build/build_ios.sh
@@ -53,27 +53,32 @@ echo short_version: $short_version
 echo pwd: `pwd`
 echo sdk_url: $sdk_url
 
-zip_name=${sdk_url##*/}
-echo zip_name: $zip_name
+unzip_name=Agora_Native_SDK_for_iOS_FULL
+sdk_url_flag=false
+if [ -z "$sdk_url" ]; then
+   sdk_url_flag=false
+   echo "sdk_url is empty"
+   echo unzip_name: $unzip_name 
+else
+   sdk_url_flag=true
+   zip_name=${sdk_url##*/}
+   echo zip_name: $zip_name
+   curl -o $zip_name $sdk_url || exit 1
+   7za x ./$zip_name -y > log.txt
+   unzip_name=`ls -S -d */ | grep Agora | sed 's/\///g'`
+   echo unzip_name: $unzip_name
+   rm -rf ./$unzip_name/bin
+   rm ./$unzip_name/commits
+   rm ./$unzip_name/package_size_report.txt
+fi
 
-curl -o $zip_name $sdk_url || exit 1
-7za x ./$zip_name -y > log.txt
-
-unzip_name=`ls -S -d */ | grep Agora | sed 's/\///g'`
-echo unzip_name: $unzip_name
-
-rm -rf ./$unzip_name/bin
-rm ./$unzip_name/commits
-rm ./$unzip_name/package_size_report.txt
-mkdir ./$unzip_name/samples
-
+mkdir -p ./$unzip_name/samples
 cp -rf ./iOS/${ios_direction} ./$unzip_name/samples/API-Example || exit 1
 ls -al ./$unzip_name/samples/API-Example/
 mv ./$unzip_name/samples/API-Example/sdk.podspec ./$unzip_name/ || exit 1
 
 echo "work space1: $WORKSPACE"
-python3 ./.github/ci/build/modify_podfile.py ./$unzip_name/samples/API-Example/Podfile || exit 1
-
+python3 ./.github/ci/build/modify_podfile.py ./$unzip_name/samples/API-Example/Podfile $sdk_url_flag || exit 1
 
 7za a -tzip result.zip -r $unzip_name > log.txt
 mv result.zip $WORKSPACE/withAPIExample_${BUILD_NUMBER}_$zip_name
